+++ CHANGELOG.md
@@ -15,6 +15,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   you to either omit `functions` for nodes, which is common for the end node,
   or specify an empty function call list, if desired.
 
+### Deprecated
+
+- The `tts` parameter in `FlowManager.__init__()` is now deprecated and will
+  be removed in a future version. The `tts_say` action now pushes a
+  `TTSSpeakFrame`.
+
 ## [0.0.17] - 2025-05-16
 
 ### Added